Convert IRNodeType enum to enum class, and rename node types.
Summary: From D1098 requested change from enum to enum class and now look at us, here we are.

Test Plan: Fixed some tests and now it should pass all old tests, this is a style change.

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1181

GitOrigin-RevId: b59ee26d351af5326b48971a9927243c34868c5e

+++ src/carnot/compiler/ir_nodes_test.cc
@@ -14,7 +14,8 @@ namespace compiler {
 
 TEST(IRTypes, types_enum_test) {
   // Quick test to make sure the enums test is inline with the type strings.
-  EXPECT_EQ(IRNodeType::number_of_types, sizeof(kIRNodeStrings) / sizeof(*kIRNodeStrings));
+  EXPECT_EQ(static_cast<int64_t>(IRNodeType::number_of_types),
+            sizeof(kIRNodeStrings) / sizeof(*kIRNodeStrings));
 }
 
 /**
